Refactor response body formatting (#99)
This commit refactors the code for formatting the response body in the
format_response function. Previously, the function assumed that the
response body was always JSON and would try to format it accordingly. Now,
the function checks the "content-type" header of the response to determine
the correct format for the body.

If the "content-type" header indicates that the body is not HTML, the body
is formatted as a JSON file instead. If the header is not present or if it
does not match the expected format for HTML, the body is still formatted as
HTML.

This change improves the robustness of the code and makes it more flexible
for different types of responses.